chore: move use-window-size to lib/hooks
