+++ src/components/tabs/hooks/index.ts
@@ -1,5 +1,5 @@
 import useOverflowRef from './use-overflow-ref'
 import useTabItems from './use-tab-items'
-import useWindowSize from './use-window-size'
+import useWindowSize from 'lib/hooks/use-window-size'
 
 export { useOverflowRef, useTabItems, useWindowSize }
